# text-mine-backend/src/main/resources/generated/api/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "text-mine",
    "version" : "v1",
    "title" : "text-mine-api"
  },
  "host" : "localhost:8080",
  "basePath" : "/",
...
  "schemes" : [ "http" ],
  "paths" : {
...
    "/articles/by-document/{documentId}" : {
      "get" : {
        "tags" : [ "articles" ],
        "summary" : "Get articles of a document",
        "description" : "",
        "operationId" : "getByDocument",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "documentId",
          "in" : "path",
          "description" : "Document Id",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "section",
          "in" : "query",
          "description" : "Section number",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "page",
          "in" : "query",
          "description" : "Page number",
          "required" : false,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/ArticleDTO"
              }
            }
          }
        }
      }
    },
...
    "/documents/attachment/{attachmentId}" : {
      "get" : {
        "tags" : [ "documents" ],
        "summary" : "Get attachment by id",
        "description" : "",
        "operationId" : "getFile",
        "produces" : [ "application/octet-stream" ],
        "parameters" : [ {
          "name" : "attachmentId",
          "in" : "path",
          "description" : "Attachment id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "type" : "string",
                "format" : "byte"
              }
            }
          }
        }
      }
    },
...
    "/notes" : {
      "get" : {
        "tags" : [ "notes" ],
        "summary" : "Get all notes",
        "description" : "",
        "operationId" : "getAll",
        "produces" : [ "application/json;charset=UTF-8" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Note"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "notes" ],
        "summary" : "Post a note",
        "description" : "",
        "operationId" : "postOne",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Note",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/NoteDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Note"
            }
          }
        }
      }
    },
    "/notes/by-document/{id}" : {
      "get" : {
        "tags" : [ "notes" ],
        "summary" : "Get note by document id",
        "description" : "",
        "operationId" : "getByDocument",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Document id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Note"
              }
            }
          }
        }
      }
    },
    "/notes/{id}" : {
      "get" : {
        "tags" : [ "notes" ],
        "summary" : "Get note by id",
        "description" : "",
        "operationId" : "getOne",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Note id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Note"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "notes" ],
        "summary" : "Update a note",
        "description" : "",
        "operationId" : "put",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "body",
          "description" : "Note",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Note"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "description" : "Note id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Note"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "notes" ],
        "summary" : "Delete a note",
        "description" : "",
        "operationId" : "delete",
        "produces" : [ "application/json;charset=UTF-8" ],
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Note id",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
...
    "ArticleDTO" : {
      "type" : "object",
      "properties" : {
        "documentId" : {
          "type" : "string"
        },
        "entryWord" : {
          "type" : "string"
        },
        "properNoun" : {
          "type" : "boolean"
        },
        "derivative" : {
          "type" : "boolean"
        },
        "editorNote" : {
          "type" : "string"
        },
        "meaning" : {
          "type" : "string"
        },
        "partOfSpeech" : {
          "type" : "array",
          "items" : {
            "type" : "string",
            "enum" : [ "PRONOMINAL_ADVERB", "PRONOUN", "NOUN", "INFINITIVE", "ADVERBIAL_PARTICIPLE", "ARTICLE", "ADVERB", "NUMERAL", "VERB", "VERB_PARTICIPLE", "PREVERB", "INTERJECTION", "CONJUNCTIVE", "ADJECTIVE", "PARTICIPLE", "MODIFIER", "SENTENTIAL", "POSTPOSITION", "POSTPOSITIONAL_ADJECTIVE", "PARTICLE", "AUXILIARY" ]
          }
        },
        "externalReferences" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "internalReferences" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "formVariants" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/FormVariant"
          }
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        }
      }
    },
...
    "EntryExample" : {
      "type" : "object",
      "properties" : {
        "exampleSentence" : {
          "type" : "string"
        },
        "page" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
    "FormVariant" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "inflections" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Inflection"
          }
        }
      }
    },
    "Inflection" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "occurrences" : {
          "type" : "integer",
          "format" : "int32"
        },
        "examples" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/EntryExample"
          }
        }
      }
    },
...
    "Note" : {
      "type" : "object",
      "required" : [ "content", "documentId", "iri", "lineRefs", "quote", "section", "subType", "type" ],
      "properties" : {
        "documentId" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "subType" : {
          "type" : "string"
        },
        "quote" : {
          "type" : "string"
        },
        "lineRefs" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "section" : {
          "type" : "integer",
          "format" : "int32"
        },
        "content" : {
          "type" : "string"
        },
        "score" : {
          "type" : "number",
          "format" : "double"
        },
        "iri" : {
          "type" : "string"
        },
        "id" : {
          "$ref" : "#/definitions/ObjectId"
        },
        "version" : {
          "type" : "integer",
          "format" : "int64"
        },
        "createdDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModifiedDate" : {
          "type" : "string",
          "format" : "date-time"
        },
        "createdBy" : {
          "type" : "string"
        },
        "lastModifiedBy" : {
          "type" : "string"
        },
        "hash" : {
          "type" : "string"
        }
      }
    },
    "NoteDTO" : {
      "type" : "object",
      "properties" : {
        "quote" : {
          "type" : "string"
        },
        "type" : {
          "type" : "string"
        },
        "subType" : {
          "type" : "string"
        },
        "content" : {
          "type" : "string"
        },
        "section" : {
          "type" : "integer",
          "format" : "int32"
        },
        "lineRefs" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "documentId" : {
          "type" : "string"
        }
      }
    },
    "ObjectId" : {
      "type" : "object",
      "properties" : {
        "timestamp" : {
          "type" : "integer",
          "format" : "int32"
        },
        "machineIdentifier" : {
          "type" : "integer",
          "format" : "int32"
        },
        "processIdentifier" : {
          "type" : "integer",
          "format" : "int32"
        },
        "counter" : {
          "type" : "integer",
          "format" : "int32"
        },
        "time" : {
          "type" : "integer",
          "format" : "int64"
        },
        "date" : {
          "type" : "string",
          "format" : "date-time"
        },
        "timeSecond" : {
          "type" : "integer",
          "format" : "int32"
        }
      }
    },
...
  }
}