# text-mine-ui/src/app/services/api/generated/model/ArticleDTO.ts
/**
 * text-mine-api
 * text-mine
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface ArticleDTO {
    documentId?: string;

    entryWord?: string;

    properNoun?: boolean;

    derivative?: boolean;

    editorNote?: string;

    meaning?: string;

    partOfSpeech?: Array<ArticleDTO.PartOfSpeechEnum>;

    externalReferences?: Array<string>;

    internalReferences?: Array<string>;

    formVariants?: Array<models.FormVariant>;

    id?: models.ObjectId;

    version?: number;

    createdDate?: Date;

    lastModifiedDate?: Date;

    createdBy?: string;

    lastModifiedBy?: string;

}
export namespace ArticleDTO {
    export enum PartOfSpeechEnum {
        PRONOMINALADVERB = <any> 'PRONOMINAL_ADVERB',
        PRONOUN = <any> 'PRONOUN',
        NOUN = <any> 'NOUN',
        INFINITIVE = <any> 'INFINITIVE',
        ADVERBIALPARTICIPLE = <any> 'ADVERBIAL_PARTICIPLE',
        ARTICLE = <any> 'ARTICLE',
        ADVERB = <any> 'ADVERB',
        NUMERAL = <any> 'NUMERAL',
        VERB = <any> 'VERB',
        VERBPARTICIPLE = <any> 'VERB_PARTICIPLE',
        PREVERB = <any> 'PREVERB',
        INTERJECTION = <any> 'INTERJECTION',
        CONJUNCTIVE = <any> 'CONJUNCTIVE',
        ADJECTIVE = <any> 'ADJECTIVE',
        PARTICIPLE = <any> 'PARTICIPLE',
        MODIFIER = <any> 'MODIFIER',
        SENTENTIAL = <any> 'SENTENTIAL',
        POSTPOSITION = <any> 'POSTPOSITION',
        POSTPOSITIONALADJECTIVE = <any> 'POSTPOSITIONAL_ADJECTIVE',
        PARTICLE = <any> 'PARTICLE',
        AUXILIARY = <any> 'AUXILIARY'
    }
}