# text-mine-ui/src/app/services/api/generated/model/FormVariant.ts
/**
 * text-mine-api
 * text-mine
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface FormVariant {
    name?: string;

    inflections?: Array<models.Inflection>;

}