# text-mine-ui/src/app/services/api/generated/model/ObjectId.ts
/**
 * text-mine-api
 * text-mine
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface ObjectId {
    timestamp?: number;

    machineIdentifier?: number;

    processIdentifier?: number;

    counter?: number;

    time?: number;

    date?: Date;

    timeSecond?: number;

}